Module Bundle.Builder

Incremental bundle builder.

Using this module it is possible to build bundle recipe incrementally. The data, that was added to the bundle will not be copied until the flush method is called.

type t

val create : unit -> t

create () creates a builder.

val put_file : ?name:string -> t -> Uri.t -> unit

put_file ?name builder uri insert a file specified by the uri. If name is specified, then the file will be stored under the specified name in the bundle.

val put_data : t -> name:string -> data:string -> unit

put_data builder ~name ~data insert data at path name.

val embed_manifest : t -> manifest -> unit

embed_manifest builder manifest embeds a manifest. If it was already embedded, then old one will be overwritten.

val flush : t -> Uri.t -> unit

flush builder output finish the building and output the resulting bundle into the file output.