Module Bap_bundle.Std

exception Not_a_bundle

type bundle

val main_bundle : unit -> bundle

main_bundle () returns a program's bundle if the program is bundled, otherwise creates a fresh new bundle in current working directory. The name of the bundle is a basename of Sys.executable_name with a bundle extension.

module Manifest : sig ... end

Program meta information.

type manifest = Manifest.t

module Bundle : sig ... end

Program Bundle.